requirements.txt
supabase
fastapi
uvicorn
jinja2
python-multipart
requests
firebase_admin
pyrebase4
python-dotenv